Split six almide-mir files at top-level boundaries to hold the line ceiling
