+++ crates/almide-mir/src/certificate_b.rs
@@ -722,198 +722,5 @@ pub fn merge_dst_i_credits(func: &MirFunction) -> usize {
         .count()
 }
 
-/// Extracted from `ownership_certificate` (codopsy8 complexity sweep, pre-scan phase 1 of
-/// 2, verbatim — the original code already scoped this as its own `{ .. }` block): a
-/// branch-MERGE dst (`Op::IfThen { dst }`) that is later RELEASED — consumed by an OUTER
-/// frame (the nested monadic-`!` chain: the inner match's merged Result moves into the
-/// outer merge) or returned — RECEIVES the arm value each arm moved in (the arm's `m`).
-/// Record that move-in as the merge object's `i` so its later `m`/`d` balances ("im", the
-/// physical rc: the arm's −1 and the merge's +1 are the same reference changing hands). An
-/// UNUSED merge dst stays event-free exactly as before. Without this the chained-`!`
-/// witness read as a bare `m` and the proven checker REJECTED it (flight-evidence-gaps F8).
-fn ownership_certificate_released_merge_dsts(
-    func: &MirFunction,
-) -> std::collections::HashSet<crate::ValueId> {
-    let mut released_merge_dsts: std::collections::HashSet<crate::ValueId> =
-        std::collections::HashSet::new();
-    let mut merge_dsts: std::collections::HashSet<crate::ValueId> = std::collections::HashSet::new();
-    for op in &func.ops {
-        match op {
-            Op::IfThen { dst: Some(d), .. } => {
-                merge_dsts.insert(*d);
-            }
-            Op::Consume { v }
-            | Op::Drop { v }
-            | Op::DropListStr { v }
-            // A RECORD/VARIANT-typed merge dst releases through the TYPED
-            // recursive drop (the #1287 record-merge seeding emits DropVariant,
-            // not Drop) — without this arm the dst never enters the released
-            // set, its IfThen `i` credit is skipped, and the DropVariant certs
-            // as a bare `d` the kernel checker rejects (unowned Dec).
-            | Op::DropVariant { v, .. } => {
-                if merge_dsts.contains(v) {
-                    released_merge_dsts.insert(*v);
-                }
-            }
-            // An INNER merge flowing out as an OUTER arm value (`Else/EndIf { val }`
-            // — the effect-TCO nested-if chain) is released the same way: the val-move
-            // rule below emits its `m`.
-            Op::Else { val: Some(v) } | Op::EndIf { val: Some(v) } => {
-                if merge_dsts.contains(v) {
-                    released_merge_dsts.insert(*v);
-                }
-            }
-            _ => {}
-        }
-    }
-    if let Some(r) = func.ret {
-        if merge_dsts.contains(&r) {
-            released_merge_dsts.insert(r);
-        }
-    }
-    // Ownership is a HEAP property: only a merge whose arm value is heap
-    // carries a reference into the dst. A SCALAR merge dst flowing out as an
-    // outer arm value (the lex-min fold's flag selects) must NOT become an
-    // object — its synthetic `i` + one-sided arm `m` certs as `i{m|}` /
-    // `i{|m}`, which the kernel-proven checker rejects while the executable
-    // verifier (scalar-blind by design) accepts — the PCC corpus-wall
-    // divergence, 2026-08-03. Same filter mirrored in `merge_dst_i_credits`.
-    let heap_objs = loop_carried_slots_heap_objs(func);
-    released_merge_dsts.retain(|d| heap_objs.contains(d));
-    released_merge_dsts
-}
 
-/// Extracted from `ownership_certificate` (codopsy8 complexity sweep, pre-scan phase 2 of
-/// 2, verbatim): the set of values EXPLICITLY moved out by an `Op::Consume` — the arm-value
-/// move for the LitStr/Var/concat arms (`lower_heap_result_arm`). Such a value's `m` is
-/// ALREADY on its object's stream, so the `Else/EndIf {val}` val-move rule in `CertScan::step`
-/// must NOT emit a SECOND `m` for it. The per-object `balance > 0` guard alone cannot catch
-/// this when the value ALIASES a still-live scope local (`else base` — the Var arm Dups
-/// base, so the shared object keeps balance 1 after the Consume, and the val-move
-/// double-`m`'d it → the `iammd` REJECT). Only the val-move-ONLY style (the effect-TCO
-/// declared-Result tail-if, whose arms never Consume) should reach the rule.
-fn ownership_certificate_consumed_values(func: &MirFunction) -> std::collections::HashSet<crate::ValueId> {
-    func.ops
-        .iter()
-        .filter_map(|op| match op {
-            Op::Consume { v } => Some(*v),
-            _ => None,
-        })
-        .collect()
-}
-
-pub fn ownership_certificate(func: &MirFunction) -> String {
-    ownership_certificate_with_poison(func).0
-}
-
-/// [`ownership_certificate`] plus whether ANY region flush emitted the
-/// always-rejecting POISON `{i|}` (a nested-region arm that cannot be
-/// represented flat). The poison REPLACES real arm events, so event COUNTS
-/// read off a poisoned certificate are meaningless — the backing gate skips
-/// them (#1146); the kernel-proven checker still rejects the poisoned cert,
-/// which is the poison's whole job.
-pub fn ownership_certificate_with_poison(func: &MirFunction) -> (String, bool) {
-    // Sequential-phase split (codopsy8 complexity sweep): the two pre-scan sets below are
-    // each an independent, self-contained computation over `func.ops` (the original code
-    // already delineated the first as its own `{ .. }` scope) — extracted verbatim as their
-    // own named functions. `CertScan::step` (protected, unchanged) and the rest of the
-    // emission pipeline below are untouched.
-    let (feeder_to_slot, slots, line_slots) = loop_carried_slots(func);
-    let depth: u32 = 0;
-    let mut s = Streams::new();
-
-    let released_merge_dsts = ownership_certificate_released_merge_dsts(func);
-    let consumed_values = ownership_certificate_consumed_values(func);
-
-    // Heap params are BORROWED (the v1 calling convention): the CALLER owns the
-    // reference and releases it, so a param contributes NO `i` event — that `+1`
-    // would be SYNTHETIC, unbacked by any runtime `Alloc`/`rc_inc` (the gate-blind
-    // use-after-free class). We still register the object identity (`of`) so that
-    // a body which releases (`Drop`/`Consume`) or returns a borrowed param WITHOUT
-    // first acquiring its own reference (a `Dup`) emits a `d`/`m` at rc 0 — which
-    // the proven checker FAULTS (REJECT), exactly the double-free that owning the
-    // caller's reference would cause. A `Dup` of the param emits the real `a`.
-    for p in &func.params {
-        if p.repr.is_heap() {
-            s.of.insert(p.value, p.value);
-        }
-    }
-
-    // Decomposed (#781, cog 123): the per-op emission lives in `CertScan::step`;
-    // the pre-scan state moved into the scan struct verbatim.
-    let mut scan = CertScan {
-        depth,
-        s,
-        released_merge_dsts,
-        consumed_values,
-        feeder_to_slot,
-        slots,
-        line_slots,
-    };
-    for op in &func.ops {
-        scan.step(op);
-    }
-
-    // Defensive: a dangling IfThen (no EndIf — malformed MIR) still flushes, so
-    // its buffered arm events land on the stream (and unbalance ⟹ reject) rather
-    // than vanish.
-    while !scan.s.frames.is_empty() {
-        scan.s.flush_branch();
-    }
-
-    // A heap return is MOVED OUT to the caller (a −1) — a move, hence `m`.
-    if let Some(r) = func.ret {
-        if scan.s.of.contains_key(&r) {
-            let o = scan.s.object_of(r);
-            scan.s.event(o, 'm');
-        }
-    }
-
-    let mut out = String::new();
-    for o in &scan.s.order {
-        out.push_str(&scan.s.stream[o]);
-        out.push('\n');
-    }
-    (out, scan.s.poisoned)
-}
-
-/// The NON-RECURRING soundness gate for the borrow-by-default calling
-/// convention, shared by the corpus classifier AND the lowering exit (#1146):
-/// EVERY `+1` event in the ownership certificate must be BACKED by a real
-/// runtime op — an `i` by an `Alloc`/`ListLit`, a heap-result call, or a
-/// credited branch merge; an `a` by a `Dup` — and every such op must have its
-/// cert line. A strict EQUALITY, so an unbacked synthetic `+1` (the
-/// gate-blind use-after-free class) AND a backed-but-uncertified op (the
-/// fs.fold_lines_chunked loop shape: one more real op than cert lines) both
-/// refuse.
-pub fn plus_one_events_backed(func: &MirFunction) -> bool {
-    let (cert, poisoned) = ownership_certificate_with_poison(func);
-    // A POISONED certificate deliberately replaced a nested-region arm's real
-    // events with the always-rejecting `{i|}` — its counts cannot be compared
-    // against the op list (the fs.fold_lines_chunked class, #1146). The
-    // poison's soundness story is the kernel checker's REJECTION of the cert
-    // itself; this equality only claims the flat-representable population.
-    if poisoned {
-        return true;
-    }
-    let i = cert.chars().filter(|c| *c == 'i').count();
-    let a = cert.chars().filter(|c| *c == 'a').count();
-    let allocs = func
-        .ops
-        .iter()
-        .filter(|o| matches!(o, crate::Op::Alloc { .. } | crate::Op::ListLit { .. }))
-        .count();
-    let heap_results = func
-        .ops
-        .iter()
-        .filter(|o| match o {
-            crate::Op::Call { dst: Some(_), result: Some(r), .. }
-            | crate::Op::CallFn { dst: Some(_), result: Some(r), .. }
-            | crate::Op::CallIndirect { dst: Some(_), result: Some(r), .. } => r.is_heap(),
-            _ => false,
-        })
-        .count();
-    let dups = func.ops.iter().filter(|o| matches!(o, crate::Op::Dup { .. })).count();
-    let merge_credits = merge_dst_i_credits(func);
-    i == allocs + heap_results + merge_credits && a == dups
-}
+include!("certificate_b_tail.rs");
