+++ crates/almide-mir/src/lower/calls.rs
@@ -770,107 +770,5 @@ fn is_admitted_effectful_entropy_env_clock(module: &str, func: &str) -> bool {
     ADMITTED_ENTROPY_ENV_CLOCK.contains(&(module, func))
 }
 
-/// Extracted from `is_admitted_effectful_pure_module_call` (codopsy8 follow-up, group 2
-/// of 3): FsRead (`read_text`/`read_bytes*`/`list_dir`/`exists`/`stat`) and FsWrite
-/// (`write`/`mkdir_p`/`remove_all`) admitted calls. Verbatim.
-fn is_admitted_effectful_fs(module: &str, func: &str) -> bool {
-    (module == "fs" && func == "read_text")
-        // `fs.read_lines` READS the filesystem — REUSES Capability::FsRead. Self-hosted
-        // as prim.read_text_file + string.lines (fs_read_lines.almd), the exact
-        // composition the native oracle is (`read_to_string(...).map(|s| s.lines()...)`).
-        || (module == "fs" && func == "read_lines")
-        || (module == "fs" && func == "read_bytes_raw")
-        || (module == "fs" && func == "list_dir")
-        || (module == "fs" && func == "read_bytes")
-        || (module == "fs" && func == "write")
-        || (module == "fs" && func == "mkdir_p")
-        // `fs.create_temp_dir` WRITES the filesystem (a mkdir under the temp
-        // root) — REUSES Capability::FsWrite, plus Entropy for the unique
-        // suffix. Self-hosted over prim.make_dir + prim.random_get
-        // (fs_create_temp_dir.almd).
-        || (module == "fs" && func == "create_temp_dir")
-        || (module == "fs" && func == "remove_all")
-        // `fs.exists` READS the filesystem (a path stat) — it REUSES Capability::FsRead
-        // (the SAME accounting as `fs.read_text`, NOT a new cap). Self-hosted to
-        // `prim.path_exists` (fs_exists.almd), so its prim floor is in the program map
-        // and the transitive cap_witness counts FsRead. UNLIKE the heap-Result fs prims,
-        // it returns a SCALAR Bool (no allocation, no scope-end drop).
-        || (module == "fs" && func == "exists")
-        // `fs.stat` READS the filesystem (the full path_filestat_get) — REUSES
-        // Capability::FsRead. Self-hosted to `prim.path_filestat` (fs_stat.almd), so its
-        // prim floor is in the program map and the transitive cap_witness counts FsRead.
-        // Returns Result[FileStat, String] (a record Ok payload).
-        || (module == "fs" && func == "stat")
-        // `fs.file_size` / `fs.modified_at` / `fs.is_dir` / `fs.is_file` READ the
-        // filesystem (each a path_filestat query) — REUSE Capability::FsRead, the
-        // fs.stat accounting. Self-hosted over prim.path_filestat (fs_file_size.almd /
-        // fs_modified_at.almd / fs_is_dir.almd — is_dir and is_file share one file).
-        || (module == "fs" && matches!(func, "file_size" | "modified_at" | "is_dir" | "is_file"))
-        // `fs.copy` READS src and WRITES dst (prim.read_text_file + prim.write_text_file,
-        // fs_copy.almd) — FsRead + FsWrite. `fs.append` is the same composition over one
-        // path (read-concat-write, fs_append.almd). `fs.create_temp_file` WRITES the
-        // fresh empty file + draws the entropy suffix (fs_create_temp_file.almd) — the
-        // fs.create_temp_dir accounting (FsWrite + Entropy).
-        || (module == "fs" && matches!(func, "copy" | "append" | "create_temp_file"))
-        // `fs.remove` READS the target's type + emptiness (path_filestat / read_dir)
-        // then WRITES the removal (remove_all floor) — FsRead + FsWrite (fs_remove.almd).
-        // `fs.write_bytes` / `fs.write_bytes_raw` WRITE the materialized byte buffer
-        // through the write floor (fs_write_bytes.almd / fs_write_bytes_raw.almd) — FsWrite.
-        // `fs.walk` / `fs.glob` recursively READ directories (read_dir + path_filestat,
-        // fs_walk.almd — glob shares the walk machinery and reads PWD, adding CliArgs).
-        || (module == "fs" && matches!(func, "remove" | "write_bytes" | "write_bytes_raw" | "walk" | "glob"))
-        // `fs.rename` WRITES the filesystem (the path_rename floor, fs_rename.almd) —
-        // FsWrite. `fs.is_symlink` READS it (the no-follow stat, fs_is_symlink.almd) — FsRead.
-        || (module == "fs" && matches!(func, "rename" | "is_symlink"))
-        // `fs.read_text_if_exists` READS the filesystem (path_filestat + the read floor,
-        // fs_read_text_if_exists.almd) — FsRead.
-        || (module == "fs" && matches!(func, "read_text_if_exists" | "read_lines_if_exists" | "read_bytes_if_exists" | "read_bytes_raw_if_exists"))
-        // `fs.fold_lines` / `fs.fold_lines_chunked` READ the filesystem — REUSE
-        // Capability::FsRead. Self-hosted typed twins over prim.read_text_file +
-        // the byte-level read_line walk (fs_fold_lines.almd); the `Map[String,
-        // Int]` accumulator routes to `_msi` in `list_heap_call_name`, any other
-        // acc to an unregistered name that walls cleanly at render. Their
-        // CLOSURE argument rides the same lift machinery every pure fold twin
-        // uses — the callback's own capabilities are captured by the lift.
-        // #1144: the ADR-0006 fallible carriers the checker rewrites
-        // `fs.fold_lines(p, z, (a, l) => g(a, l)!)` / `fs.for_each_line(p, (l)
-        // => g(l)!)` into — same read, same capability, only the callback's
-        // answer changed shape. BOTH must be admitted, including
-        // `__fallible_for_each_line`, which has no wasm twin: an unadmitted
-        // effectful call is refused HERE, in lower_function, which walls the
-        // enclosing REAL function and trips the walled-real ratchet
-        // (proofs/walled-real-baseline.txt is permanently 0). Admitting it
-        // keeps lowering TOTAL and moves the honest refusal to the renderer,
-        // where the unlinked twin name lives — the same place the `_x`
-        // accumulator cells of the total family wall.
-        || (module == "fs" && matches!(func, "fold_lines" | "fold_lines_chunked" | "fold_lines_range" | "__fallible_fold_lines" | "__fallible_for_each_line"))
-}
 
-/// Extracted from `is_admitted_effectful_pure_module_call` (codopsy8 follow-up, group 3
-/// of 3): Stdout (`io.print`) and Stdin (`io.read_line`/`io.read_n_bytes`) admitted
-/// calls. Verbatim.
-fn is_admitted_effectful_io(module: &str, func: &str) -> bool {
-    (module == "io" && func == "print")
-        || (module == "io" && func == "read_line")
-        // `io.read_n_bytes` READS standard input (the SIBLING of read_line) — REUSES
-        // Capability::Stdin. Self-hosted to `prim.read_n_bytes` (io_read_n_bytes.almd → the
-        // WASI fd-0 $read_n_bytes floor), so its prim is in the program map and the transitive
-        // cap_witness counts Stdin. Returns a heap Bytes block (flat Drop, no nested handles).
-        || (module == "io" && func == "read_n_bytes")
-        // `io.write` (Bytes) / `io.write_bytes` (List[Int]) WRITE standard output — the
-        // SAME Capability::Stdout as `io.print`, over the same single-iovec fd_write floor
-        // (io_write.almd → prim.fd_write), so their prims are in the program map and the
-        // transitive cap_witness counts Stdout. Both return Unit.
-        || (module == "io" && matches!(func, "write" | "write_bytes"))
-        // `io.read_all` READS standard input to EOF — REUSES Capability::Stdin. Self-hosted
-        // as a chunked `io.read_n_bytes` loop (io_read_all.almd, #876), so its transitive
-        // cap_witness reaches the same prim.read_n_bytes floor and counts Stdin. Returns an
-        // owned String (flat Drop).
-        || (module == "io" && func == "read_all")
-        // `io.read_byte` READS one byte of standard input — REUSES Capability::Stdin.
-        // Self-hosted over the same prim.read_n_bytes floor (io_read_byte.almd),
-        // returning the byte 0..255 or -1 on EOF (a SCALAR Int, no ownership).
-        || (module == "io" && func == "read_byte")
-}
-
-include!("calls_b.rs");
+include!("calls_tail.rs");
